+++ Utils/ItemStorage.lua
@@ -9,27 +9,46 @@ local Storage = {}
 addon.ItemStorage = Storage
 
 local StoredItems = {}
+local private = {}
 
+--[[
+   Each entry index marks the type of item, and can have following fields:
+      'bagged' - Must be in player's bags to be initialized on login?
+]]
 Storage.AcceptedTypes = {
-   ["to_trade"]    = true, -- Items that should be traded to another player
-   ["award_later"] = true, -- Items that should be used in a later session
-   ["other"]       = true, -- Unspecified
+   ["to_trade"]    = { -- Items that should be traded to another player
+      bagged = true,
+   },
+   ["award_later"] = { -- Items that should be used in a later session
+      bagged = false,
+   },
+   ["temp"]        = { -- Items we're temporarily storing
+      bagged = true
+   },
+   ["other"]       = { -- Unspecified
+      bagged = false
+   },
 }
 
-local item_prototype = {
+-- Item Class:
+local item_class = {
    type = "other", -- Default to unspecified
    time_remaining = 0, -- NOTE For now I rely on this not being updated for timeout checks. It should be precise enough, but needs testing
    time_added = 0,
-   exists = false,
+   inBags = false,
    link = "",
    args = {}, -- User args
-}
 
-local item_methods = {
-   GetTimeRemaining = function(self)
-      return self.time_remaining
+   Store = function(self)
+      tinsert(db.itemStorage, self)
+      return self
+   end,
+   Unstore = function(self)
+      tDeleteItem(db.itemStorage, self)
+      return self
    end,
 }
+
 -- lua
 local error, table, tostring, tinsert, tremove, type, select, FindInTableIf, time, tFilter, setmetatable, CopyTable, tDeleteItem, ipairs
     = error, table, tostring, tinsert, tremove, type, select, FindInTableIf, time, tFilter, setmetatable, CopyTable, tDeleteItem, ipairs
@@ -40,102 +59,71 @@ function addon:InitItemStorage()-- Extract items from our SV. Could be more eleg
    db = self:Getdb()
    local Item;
    for k, v in ipairs(db.itemStorage) do
-      Item = Storage:StoreItem(v.link, v.type, "restored", v)
-      if not Item.exists then -- Item probably no longer exists?
-         addon:Debug("Error - ItemStorage, couldn't add db item:", v.link)
-         local key = FindInTableIf(db.itemStorage, function(d) return addon:ItemIsItem(v.link, d.link) end)
-         if key then
-            tremove(db.itemStorage, key)
-         else
-            addon:Debug("Error - Unable to remove item from db?!")
-         end
+      Item = Storage:New(v.link, v.type, "restored", v)
+      if not Item.inBags and Storage.AcceptedTypes[Item.type] then -- Item probably no longer exists?
+         addon:Debug("Error - ItemStorage, db item no longer in bags", v.link)
+         Storage:RemoveItem(Item)
       end
    end
 end
 
-local function findItemInBags(link)
-   addon:DebugLog("Storage: searching for item:",link)
-   if link and link ~= "" then
-      for container=0, _G.NUM_BAG_SLOTS do
-   		for slot=1, GetContainerNumSlots(container) or 0 do
-            if addon:ItemIsItem(link, GetContainerItemLink(container, slot)) then
-               addon:DebugLog("Found item at",container, slot)
-               return container, slot
-            end
-         end
-      end
-   end
-   addon:DebugLog("Error - Couldn't find item")
-end
-
-local function newItem(link, type, time_remaining)
-   local Item = setmetatable({}, {
-      __index = item_prototype,
-      __tostring = function(self)
-         return self.link
-      end,
-   })
-   Item.link = link
-   Item.type = type and type or Item.type
-   Item.time_remaining = time_remaining and time_remaining or Item.time_remaining
-   Item.time_added = time()
-   return Item
-end
-
---- Pesistantly store an item
--- Item is stored in Ace3 db and an internal db for future use. The internal item object is returned
+--- Initiates a new item of item_class
 -- @param item ItemLink|ItemString|ItemID of the item
--- @param type The storage type. Used for different handlers, @see Storage.AcceptedTypes
+-- @param type Optional type for used in various functions, @see Storage.AcceptedTypes
 -- @param ... Userdata stored in the returned 'Item.args'. Directly stored if provided as table, otherwise as '{...}'.
--- @returns Item @see 'item_prototype' when the item is stored succesfully. Has flag Item.exists if present in bags.
-function Storage:StoreItem(item, typex, ...)
+-- @returns Item @see 'item_class' when the item is stored succesfully. Has flag Item.inBags if present in bags.
+function Storage:New(item, typex, ...)
    if not typex then typex = "other" end
    if not self.AcceptedTypes[typex] then error("Type: " .. tostring(typex) .. " is not accepted. Accepted types are: " .. table.concat(self.AcceptedTypes, ", "),2) end
-   addon:Debug("Storage:StoreItem",item,typex,...)
-   local c,s = findItemInBags(item)
+   addon:Debug("Storage:New",item,typex,...)
+   local c,s,time_remaining = private:findItemInBags(item)
    local Item
    if not (c and s) then
       -- The Item is not in our bags
-      Item = newItem(item, typex)
+      Item = private:newItem(item, typex)
    else
-      local time_remaining = addon:GetContainerItemTradeTimeRemaining(c,s)
-      Item = newItem(item, typex, time_remaining)
-      Item.exists = true -- The item is in our bags
+      Item = private:newItem(item, typex, time_remaining)
+      Item.inBags = true -- The item is in our bags
    end
-   if select(1, ...) == "restored" then
+   if select(1, ...) == "restored" then -- Special case, gets stored
       local OldItem = select(2, ...)
       Item.time_added = OldItem.time_added -- Restore original time added
       Item.args = OldItem.args
-   else -- We need to store it in db as well
-      Item.args = type(...) == "table" and ... or {...}
-      tinsert(db.itemStorage, Item)
+   else
+      Item.args = ... and type(...) == "table" and ... or {...}
    end
    tinsert(StoredItems, Item)
    return Item
 end
 
 --- Remove an item from storage
--- @param item Either an itemlink (@see GetItemInfo) or the Item object returned by :StoreItem
+-- @param item Either an itemlink (@see GetItemInfo) or the Item object returned by :New
 -- @return True if successful
 function Storage:RemoveItem(item)
    addon:Debug("Storage:RemoveItem", item)
-   if type(item) == "table" then -- Our Item object
-      if StoredItems[item] and db.itemStorage[item] then
-         tDeleteItem(StoredItems, item)
-         tDeleteItem(db.itemStorage, item)
-         return true
-      else -- Item didn't exist, try to extract itemlink
-         return self:RemoveItem(item.link)
+   local item_link
+   if type(item) == "table" then -- Maybe our Item object
+      if not (item.type and item.link) then -- Nope
+         return error("Item `item` is not the correct class", 2)
       end
+      item_link = item.link
    elseif type(item) == "string" then -- item link (hopefully)
-      local key = FindInTableIf(StoredItems, function(v) return addon:ItemIsItem(v.link,item) end)
-      if key then
-         tremove(StoredItems, key)
-         tremove(db.itemStorage, key)
-         return true
-      end
+      item_link = item
+   else
+      return error("Unknown item")
    end
-   addon:Debug("Error - Couldn't remove item")
+   -- Find and delete the item
+   local key1 = private:FindItemInTable(db.itemStorage, item_link)
+   local key2 = private:FindItemInTable(StoredItems, item_link)
+   if key1 then tremove(db.itemStorage, key1) end
+   if key2 then tremove(StoredItems, key2) end
+
+   if not (key1 and key2) then
+      addon:Debug("Error - Couldn't remove item", key1, key2)
+      return false
+   else
+      return true
+   end
 end
 
 --- Removes all items of a specific type
@@ -145,9 +133,7 @@ function Storage:RemoveAllItemsOfType(type)
    -- Do it in reverse for speed
    for i = #StoredItems, 1, -1 do
       if StoredItems[i].type == type then
-         -- REVIEW Are we really guaranteed to have the same index in both tables?
-         tremove(StoredItems, i)
-         tremove(db.itemStorage, i)
+         self:RemoveItem(StoredItems[i])
       end
    end
 end
@@ -164,7 +150,7 @@ end
 -- @return The Item object, or nil if not found
 function Storage:GetItem(item)
    if type(item) ~= "string" then return error("'item' is not a string/ItemLink", 2) end
-   return select(2, FindInTableIf(StoredItems, function(v) return addon:ItemIsItem(v.link, item) end))
+   return select(2, private:FindItemInTable(StoredItems, item))
 end
 
 --- Returns all stored Items of a specific type
@@ -203,13 +189,54 @@ function Storage:GetAllItemsMultiPred(...)
       end, true)
 end
 
---- Returns Container and Slot ids of the item.
+--- Returns Container and Slot ids of an item.
 -- @param item The item to find slots for, either 'Item' object or ItemLink.
--- @return container,slot The position of the item in the player's bags.
+-- @return container,slot,time_remaining The position of the item in the player's bags and it's remaining trade time (if present).
 function Storage:GetItemContainerSlot (item)
    if type(item) == "table" then -- Our Item object
-      return findItemInBags(item.link)
+      return private:findItemInBags(item.link)
    elseif type(item) == "string" then
-      return findItemInBags(item)
+      return private:findItemInBags(item)
    end
+end
+
+function private:FindItemInTable(table, item1)
+   return FindInTableIf(table, function(item2) return addon:ItemIsItem(item1, item2.link) end)
+end
+
+function private:findItemInBags(link)
+   addon:DebugLog("Storage: searching for item:",link)
+   if link and link ~= "" then
+      local c,s,t
+      for container=0, _G.NUM_BAG_SLOTS do
+   		for slot=1, GetContainerNumSlots(container) or 0 do
+            if addon:ItemIsItem(link, GetContainerItemLink(container, slot)) then -- We found it
+               addon:DebugLog("Found item at",container, slot)
+               c, s = container, slot
+               -- Now we need to ensure we don't have multiple of it
+               t = addon:GetContainerItemTradeTimeRemaining(c,s)
+               if t > 0 then
+                  -- if the item is tradeable, then we've most likely just received it, and can safely return
+                  return c,s,t
+               end
+            end
+         end
+      end
+      addon:DebugLog("Error - Couldn't find item")
+      return c,s,t
+   end
+end
+
+function private:newItem(link, type, time_remaining)
+   local Item = setmetatable({}, {
+      __index = item_class,
+      __tostring = function(self)
+         return self.link
+      end,
+   })
+   Item.link = link
+   Item.type = type or Item.type
+   Item.time_remaining = time_remaining or Item.time_remaining
+   Item.time_added = time()
+   return Item
 end