+++ changelog.md
@@ -1,3 +1,22 @@
+### v2.11.0
+---
+* **Loot Status**  
+* The loot status now registers all items dropped by a boss, even if the response indicating if it's tradeable isn't received from the candidate.
+* Implemented code to ignore certain loot sources and types (particularly Opulence trash piles), which should overall make the loot status more reliable.
+
+
+###### Bugfixes  
+* *The TradeUI is now able to handle multiple instances of the same item.*
+* *More error correcting code for corrupted loot history*
+* *No longer logs `non_tradeable` loots when the addon isn't being used, i.e. in dungeons.*
+* *Mostly fixed the worst texture stretching throughout the UI.*
+
+
+###### Dev
+* Added `addon.Compat` namespace for handling backwards compatibility.
+* Added `addon.lootGUIDToIgnore` table indicating guid's that won't have their loot registered.
+
+
 ### v2.10.3
 ---
 ###### Bugfixes